Got the desired interface down for the CameraActivity. Added any necessary listeners and added enabled and disabled functionality when such actions occur (when capture button is pressed, show two icons in the middle)
Added two buttons that will approve and disapprove captured picture.
Added listener related to TextureView that will listen to any pressed event. Will be used for manual autofocus later. Also ensured it was disabled after when the user presses the capture button.

Added that took into account of the freezing issue when the user resumes the activity of the camera from onPause() (when the screen is off and then the user turns it back on). Also compiled it into one method and then reused it for certain areas of the code to stop repetition of code.